rename report2.py to report.py, superseding the old file by that name, and remove summarize.py
